Refactoring current tests
